+++ src/commands/fetch.ts
@@ -1,5 +1,10 @@
 import { Message } from "revolt.js";
-import { styles, login, fetchChannel } from "../lib/constants/index.js";
+import {
+  styles,
+  login,
+  renderMessages,
+  fetchChannel,
+} from "../lib/constants/index.js";
 
 const fetchMsgs = async function (
   userType: string,
@@ -15,7 +20,7 @@ const fetchMsgs = async function (
         "You need to specify whether the account is a user or a bot (--user/--bot), a token, a channel ID, the number of messages to fetch (max 100) and optionally a custom API URL (all in quotes)."
       )
     );
-    return process.exit();
+    return process.exit(1);
   }
   if (!token) {
     console.log(
@@ -23,7 +28,7 @@ const fetchMsgs = async function (
         "You need to specify a token, a channel ID, the number of messages to fetch (max 100) and optionally a custom API URL (all in quotes)."
       )
     );
-    return process.exit();
+    return process.exit(1);
   }
   if (!channel) {
     console.log(
@@ -31,7 +36,7 @@ const fetchMsgs = async function (
         "You need to specify a channel ID, the number of messages to fetch (max 100) and optionally a custom API URL (all in quotes)."
       )
     );
-    return process.exit();
+    return process.exit(1);
   }
   if (!rawAmount) {
     console.log(
@@ -39,7 +44,7 @@ const fetchMsgs = async function (
         "You need to specify the number of messages to fetch (max 100) and optionally a custom API URL (all in quotes)."
       )
     );
-    return process.exit();
+    return process.exit(1);
   }
 
   // log in
@@ -53,36 +58,12 @@ const fetchMsgs = async function (
     // send the message
     try {
       const amount = parseInt(rawAmount);
-      const msgs = await channel2!.fetchMessages({
-        limit: amount,
-        sort: "Latest",
-      });
-      if (!msgs.length) {
-        console.log(styles.error("There aren't any messages to archive!"));
-        process.kill(process.pid);
+      const output = await renderMessages(channel2, true, amount);
+      if (output === "noMsgs") {
+        console.log(styles.error("There aren't any messages to display!"));
+        return process.exit(1);
       }
-      console.log(styles.success("Sucessfully fetched messages!"));
-      let rawoutput = [""];
-      function contentToSend(msg: Message) {
-        if (msg.content === null) {
-          if (msg.system === null) {
-            return "No content.";
-          } else {
-            return msg.system?.type;
-          }
-        } else {
-          return msg.content;
-        }
-      }
-      msgs.forEach((msg) => {
-        const content = contentToSend(msg);
-        const obj = `${styles.info(
-          `${msg.author?.username} (${msg.author_id}) - ${msg._id} (${msg.createdAt})`
-        )}\n${content}`;
-        rawoutput.push(obj);
-      });
-      const output =
-        styles.header("Messages (newest first):\n") + rawoutput.join("\n");
+
       console.log(output);
     } catch (error) {
       console.log(
@@ -92,8 +73,8 @@ const fetchMsgs = async function (
       );
     }
 
-    // for SOME reason we need to end the process manually after sending the message - is something lingering?
-    process.kill(process.pid);
+    // for SOME reason we need to end the process manually - is something lingering?
+    return process.exit(0);
   });
 };