+++ src/types/Interactive.d.ts
@@ -1,0 +1,4 @@
+export type NameIDPair = {
+  name: string;
+  id: string;
+};